Update index.ts
changed the server listening port, so the following works.
Now with the changes, if you have another system, where you host the server code, and you try to access it, with this change, it will work, because, by default, http uses port 80 by default as said by @NihalNavath in his commits.
And, you can still self host it in the same computer, just use localhost:80 instead of localhost:8080 like we did before.