The barrels return + crates spawn random things
After 4-5 months of asking you about barrels, I finally figured out, how to do it, and here we are, with a barrels update
Now, the crates and the soviet crate all spawn random guns! instead of hardcoding things, it now works, with a local gun list, that needs to be updated, if it should be added to that type of crate.